designing and some finisg=hing works like pagination validation missing decorators

+++ users/views.py
@@ -1,13 +1,34 @@
 from django.shortcuts import render, redirect
-from .forms import SignupForm, LoginForm, UserRegisterForm, OfferForm
-from .models import CustomUser
+from .forms import (
+    SignupForm,
+    LoginForm,
+    UserRegisterForm,
+    OfferForm,
+    ComplaintForm,
+    ComplaintRespondForm,
+)
+
+# from .models import CustomUser
 from django.contrib import messages
 from django.contrib.auth import authenticate, login
 from django.shortcuts import get_object_or_404
 from users.utils import create_notification
-from users.models import Notification, Offers
-from users.utils import create_offers
-from shop.models import Accessories, BookingService, BookingFood, BookingAccessories
+from users.models import Notification, Offers, Complaints, CustomUser
+from django.views.generic import ListView
+
+# from users.utils import create_offers
+from shop.models import (
+    Accessories,
+    BookingService,
+    BookingFood,
+    BookingAccessories,
+    Appointment,
+    ServiceOffered,
+    DogFood,
+    
+)
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.mixins import LoginRequiredMixin
 
 
 def Signup(request):
@@ -17,7 +38,7 @@ def Signup(request):
             user = form.save()
             login(request, user)
             messages.success(
-                request, "You Account Created Succesfully ", extra_tags="alert-success"
+                request, "You Account Created Successfully ", extra_tags="alert-success"
             )
             return redirect("login")
         else:
@@ -55,6 +76,7 @@ def log_in(request):
     return render(request, "login.html", context)
 
 
+@login_required(login_url="signup")
 def home(request):
     if request.user.is_authenticated:
         create_notification(request.user, "Welcome Our App")
@@ -69,6 +91,7 @@ def log_out(request):
     return redirect("signup")
 
 
+@login_required(login_url="signup")
 def dashboard(request):
     if request.user.role == "SHOP OWNER":
         return redirect("shop_dashboard")
@@ -76,55 +99,86 @@ def dashboard(request):
         return redirect("users_dashboard")
 
 
+@login_required(login_url="signup")
 def Shop_owner_dashboard(request):
-    user_count = CustomUser.objects.count()
-    role_count = CustomUser.objects.filter(role="SHOP OWNER").count()
-    accessories = Accessories.objects.all()  # Fetch all accessories
+    if request.user.role == "SHOP OWNER":
+        user_count = CustomUser.objects.count()
+        role_count = CustomUser.objects.filter(role="SHOP OWNER").count()
+        accessories = Accessories.objects.all().count() # Fetch all accessories
 
-    context = {
-        "page_title": "Shop Owner",
-        "user_count": user_count,
-        "role_count": role_count,
-        "all_accessorie": accessories,
-    }
-    return render(request, "shop_owner_dashboard.html", context)
+        context = {
+            "page_title": "Shop Owner",
+            "user_count": user_count,
+            "role_count": role_count,
+            "all_offers":Offers.objects.all().count(),
+            "all_accessorie": accessories,
+            "all_foods":DogFood.objects.all().count(),
+            "all_service":ServiceOffered.objects.all().count(),
+            "all_users": BookingService.objects.all().count(),
+            "all_accessories": BookingAccessories.objects.all().count(),
+            "all_food": BookingFood.objects.all().count(),
+            "appointments": Appointment.objects.all().count(),
+            "all_complaints":Complaints.objects.all().count(),
+        }
+        return render(request, "shop_owner_dashboard.html", context)
+    return redirect("403")
 
 
+@login_required(login_url="signup")
 def users_dashboard(request):
     context = {
         "page_title": "User DashBoard",
         "all_users": BookingService.objects.all(),
         "all_accessories": BookingAccessories.objects.all(),
         "all_foods": BookingFood.objects.all(),
+        "appointments": Appointment.objects.filter(username=request.user),
     }
 
     return render(request, "users_dashboard.html", context)
 
 
-def view_users(request):
-    context = {"page_title": "All Users", "all_users": CustomUser.objects.all()}
-    return render(request, "view_users.html", context)
+class UserListView(LoginRequiredMixin, ListView):
+    model = CustomUser
+    template_name = "view_users.html"
+    context_object_name = "all_users"
+    paginate_by = 8
+    ordering = ["id"]
+
+    def dispatch(self, request, *args, **kwargs):
+        if request.user.role != "SHOP OWNER":
+            return redirect("403")
+        return super().dispatch(request, *args, **kwargs)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["page_title"] = "All Registered User"
+        return context
 
 
+@login_required(login_url="signup")
 def add_users(request):
-    if request.method == "POST":
-        form = UserRegisterForm(request.POST)
-        if form.is_valid():
-            form.save()
-            messages.success(
-                request, "Registered user successfully", extra_tags="alert-success"
-            )
-            return redirect("shop_dashboard")
-        else:
-            for error_list in form.errors.values():
-                for errors in error_list:
-                    messages.error(request, errors, extra_tags="alert-danger")
+    if request.user.role == "SHOP OWNER":
 
-    context = {"page_title": "Register User", "form": UserRegisterForm()}
+        if request.method == "POST":
+            form = UserRegisterForm(request.POST)
+            if form.is_valid():
+                form.save()
+                messages.success(
+                    request, "Registered user successfully", extra_tags="alert-success"
+                )
+                return redirect("shop_dashboard")
+            else:
+                for error_list in form.errors.values():
+                    for errors in error_list:
+                        messages.error(request, errors, extra_tags="alert-danger")
 
-    return render(request, "add_update_users.html", context)
+        context = {"page_title": "Register User", "form": UserRegisterForm()}
+
+        return render(request, "add_update_users.html", context)
+    return redirect("403")
 
 
+@login_required(login_url="signup")
 def update_users(request, pk):
     to_update = get_object_or_404(CustomUser, id=pk)
     if request.method == "POST":
@@ -147,15 +201,21 @@ def update_users(request, pk):
     return render(request, "add_update_users.html", context)
 
 
+@login_required(login_url="signup")
 def delete_users(request, pk):
-    to_delete = get_object_or_404(CustomUser, id=pk)
-    to_delete.delete()
-    messages.success(request, "Users Deleted Successfully", extra_tags="alert-success")
+    if request.user.role == "SHOP OWNER":
+        to_delete = get_object_or_404(CustomUser, id=pk)
+        to_delete.delete()
+        messages.success(
+            request, "Users Deleted Successfully", extra_tags="alert-success"
+        )
 
-    return redirect("view_users")
+        return redirect("view_users")
+    return redirect("403")
 
 
 # notification
+@login_required(login_url="signup")
 def view_notification(request):
     context = {
         "page_title": "View Notification",
@@ -164,6 +224,7 @@ def view_notification(request):
     return render(request, "notification.html", context)
 
 
+@login_required(login_url="signup")
 def read_offers(request, pk):
     notification = get_object_or_404(Notification, id=pk)
     notification.is_read = True
@@ -172,6 +233,7 @@ def read_offers(request, pk):
 
 
 # offers
+@login_required(login_url="signup")
 def view_offers(request):
     context = {
         "page_title": "All Offers",
@@ -182,69 +244,149 @@ def view_offers(request):
     return render(request, "add_view_offers.html", context)
 
 
-# def add_offers(request):
-#     if request.method == "POST":
-#         form = OfferForm(request.POST)
-#         if form.is_valid():
-#             offers = form.save(commit=False)
-#             offers.username = request.user
-#             form.save()
-
-#             messages.success(
-#                 request, "Offers Added  Successfully", extra_tags="alert-success"
-#             )
-#             notify=CustomUser.objects.filter(is_active=True)
-#             offer=offers.save
-
-#             if request.user.is_authenticated:
-#                 create_offers(notify,offer)
-#             return redirect("shop_dashboard")
-
-
-#         else:
-#             for error_list in form.errors.values():
-#                 for errors in error_list:
-#                     messages.success(request, errors, extra_tags="alert-danger")
-#     context = {"page_title": "View Offers", "form": OfferForm()}
-#     return render(request, "add_view_offers.html", context)
+@login_required(login_url="signup")
 def add_offers(request):
+    if request.user.role == "SHOP OWNER":
+        if request.method == "POST":
+            form = OfferForm(request.POST)
+            if form.is_valid():
+                offers = form.save(commit=False)
+                offers.username = (
+                    request.user
+                )  # Ensure this field exists in the Offers model
+                offers.save()
+
+                messages.success(
+                    request, "Offers Added Successfully", extra_tags="alert-success"
+                )
+                return redirect("shop_dashboard")
+            else:
+
+                for error_list in form.errors.values():
+                    for error in error_list:
+                        messages.error(request, error, extra_tags="alert-danger")
+
+        context = {"page_title": "View Offers", "form": form}
+        return render(request, "add_view_offers.html", context)
+    return redirect("403")
+
+
+@login_required(login_url="signup")
+def delete_offers(request, pk):
+    if request.user.role == "SHOP OWNER":
+        to_delete = get_object_or_404(Offers, id=pk)
+        to_delete.delete()
+        messages.success(
+            request, "Offers Deleted Successfully", extra_tags="alert-success"
+        )
+        return redirect("shop_dashboard")
+    return redirect("403")
+
+
+@login_required(login_url="signup")
+def unauthorized(request):
+    context = {"page_title": "403"}
+    return render(request, "404.html", context)
+
+
+class ComplaintListView(LoginRequiredMixin, ListView):
+    model = Complaints
+    template_name = "view_complaints.html"
+    context_object_name = "all_bookings"
+    paginate_by = 6
+    ordering = ["-id"]
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["page_title"] = "View Complaints"
+        context["all_bookings"] = Complaints.objects.filter(username=self.request.user)
+        return context
+
+
+@login_required(login_url="signup")
+def add_complaints(request):
     if request.method == "POST":
-        form = OfferForm(request.POST)
+        form = ComplaintForm(request.POST)
         if form.is_valid():
-            offers = form.save(commit=False)
-            offers.username = (
-                request.user
-            )  # Ensure this field exists in the Offers model
-            offers.save()
+            complaints = form.save(commit=False)
+            complaints.username = request.user
+            complaints.save()
+            messages.success(
+                request, "Complaint Registered Successfully", extra_tags="alert-success"
+            )
+            return redirect("view_complaints")
+        else:
+            for error_list in form.errors.values():
+                for errors in error_list:
+                    messages.error(request, errors, extra_tags="alert-danger")
+    context = {"page_title": "Register complaint", "form": ComplaintForm}
+    return render(request, "add_update_users.html", context)
+
+
+@login_required(login_url="signup")
+def update_complaints(request, pk):
+    to_update = get_object_or_404(Complaints, id=pk)
+    if request.method == "POST":
+        form = ComplaintForm(request.POST, instance=to_update)
+        if form.is_valid():
+            form.save()
 
             messages.success(
-                request, "Offers Added Successfully", extra_tags="alert-success"
+                request, "Complaint Updated Successfully", extra_tags="alert-success"
+            )
+            return redirect("view_complaints")
+        else:
+            for error_list in form.errors.values():
+                for errors in error_list:
+                    messages.error(request, errors, extra_tags="alert-danger")
+    context = {
+        "page_title": "Updated complaint",
+        "form": ComplaintForm(instance=to_update),
+    }
+    return render(request, "add_update_users.html", context)
+
+
+@login_required(login_url="signup")
+def delete_complaints(request, pk):
+    to_delete = get_object_or_404(Complaints, id=pk)
+    to_delete.delete()
+    messages.success(
+        request, "Complaint Deleted Successfully", extra_tags="alert-success"
+    )
+    return redirect("view_complaints")
+
+
+@login_required(login_url="signup")
+def respond(request, pk):
+    to_update = get_object_or_404(Complaints, id=pk)
+    if request.method == "POST":
+        form = ComplaintRespondForm(request.POST, instance=to_update)
+        if form.is_valid():
+            form.save()
+
+            messages.success(
+                request, "Complaint Updated Successfully", extra_tags="alert-success"
             )
             return redirect("shop_dashboard")
         else:
-            # Print out form errors for debugging
-            print(form.errors)  # Check the console for any validation issues
             for error_list in form.errors.values():
-                for error in error_list:
-                    messages.error(
-                        request, error, extra_tags="alert-danger"
-                    )  # Change to error
-
-    else:
-        form = OfferForm()  # Initialize a new form for GET requests
-
-    context = {"page_title": "View Offers", "form": form}
-    return render(request, "add_view_offers.html", context)
+                for errors in error_list:
+                    messages.error(request, errors, extra_tags="alert-danger")
+    context = {
+        "page_title": "Complaint Responds",
+        "form": ComplaintRespondForm(instance=to_update),
+    }
+    return render(request, "add_update_users.html", context)
 
 
-def delete_offers(request, pk):
-    to_delete = get_object_or_404(Offers, id=pk)
-    to_delete.delete()
-    messages.success(request, "Offers Deleted Successfully", extra_tags="alert-success")
-    return redirect("shop_dashboard")
+class AllComplaintListView(LoginRequiredMixin, ListView):
+    model = Complaints
+    template_name = "all_complaints.html"
+    context_object_name = "all_bookings"
+    paginate_by = 6
+    ordering = ["-id"]
 
-
-
-def unauthorized(request):
-    context={"page_title":"403"}
-    return render(request,"404.html",context)
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["page_title"] = "All Complaints"
+        return context
